feat: Implement int8 QAT and static quantization for inference
- Added `QuantStub` and `DeQuantStub` across key architectural components (`Encoder`, `Generator`, `Discriminator`, `Hyperprior`).
- Replaced non-quantizable tensor operations (e.g., `torch.cat`, `torch.add`) with `nn.quantized.FloatFunctional` equivalents.
- Updated `ChannelNorm2D` wrapper to handle both standard continuous flows and intermediate dequantization-requantization paths.
- Adjusted `utils.load_model` to properly inject quantization structures *before* loading state_dicts, ensuring correct tracking of learned observer scales/zero-points.
- Fixed optimizer re-initialization in `train.py` for QAT models.
- Added `--quantize` flags in `train.py` and `compress.py` pipelines to toggle quantization aware training and static quantization inference respectively.

+++ src/compression/hyperprior_model.py
@@ -310,7 +310,7 @@ class HyperpriorDensity(nn.Module):
         x:  The values at which to evaluate the cumulative densities.
             torch.Tensor - shape `(C, 1, *)`.
         """
-        logits = x
+        logits = x.float()  # keep FP32 for numerical stability regardless of input dtype
 
         for k in range(len(self.filters)+1):
             H_k = getattr(self, 'H_{}'.format(str(k)))  # Weight
